Refactor main.py to reduce complexity of it
